add app history view and nav for app detail
