+++ frontend/less/admin.less
@@ -135,3 +135,44 @@ table.table tr > td.org_tags {
 		color: #fff;
 	}
 }
+
+.app-detail-nav-bar {
+	overflow: auto;
+	position: relative;
+	margin-top: 1em;
+}
+
+.app-detail-nav-bar ul {
+	padding: 0;
+}
+.app-detail-nav-bar li {
+	display: inline-block;
+	a {
+		display: inline-block;
+		padding: 1em 1em 0em 1em;
+	}
+}
+
+.app-detail-nav-bar li.active a {
+	color: black;
+}
+.app-detail-nav-bar hr {
+	margin-top: 0.75em;
+	margin-bottom: 0;
+	border-top-width: 3px;
+}
+
+.update-status-button {
+	position: absolute;
+	right: 0;
+	top: 0;
+}
+
+.action_bar h1 {
+	overflow: auto;
+	position: relative;
+}
+.action_bar a.btn {
+	position: absolute;
+	right: 0;
+}
