+++ frontend/less/main.less
@@ -15,6 +15,7 @@
 @import "../../node_modules/bootstrap/less/tables.less";
 @import "../../node_modules/bootstrap/less/forms.less";
 @import "../../node_modules/bootstrap/less/buttons.less";
+@import "../../node_modules/bootstrap/less/navs.less";
 @import "../../node_modules/bootstrap/less/pagination.less";
 
 // Utility classes
