+++ intake/tests/views/test_admin_views.py
@@ -1,4 +1,3 @@
-
 from unittest import skipUnless
 from unittest.mock import patch
 
@@ -12,92 +11,6 @@ from intake.tests.base_testcases import IntakeDataTestCase, DELUXE_TEST
 from project.jinja2 import url_with_ids
 
 import intake.services.bundles as BundlesService
-
-
-class TestApplicationDetail(IntakeDataTestCase):
-
-    fixtures = [
-        'counties',
-        'organizations', 'mock_profiles',
-        'mock_2_submissions_to_a_pubdef',
-        'mock_2_submissions_to_sf_pubdef',
-        'mock_1_submission_to_multiple_orgs', 'template_options'
-        ]
-
-    def get_detail(self, submission):
-        url = reverse(
-            'intake-app_detail', kwargs=dict(submission_id=submission.id))
-        result = self.client.get(url)
-        return result
-
-    def assertHasDisplayData(self, response, submission):
-        for field, value in submission.answers.items():
-            if field in self.display_field_checks:
-                escaped_value = html_utils.conditional_escape(value)
-                self.assertContains(response, escaped_value)
-
-    @patch('intake.notifications.slack_submissions_viewed.send')
-    def test_logged_in_user_can_get_submission_display(self, slack):
-        self.be_apubdef_user()
-        submission = self.a_pubdef_submissions[0]
-        response = self.get_detail(submission)
-        self.assertEqual(response.context_data['form'].submission, submission)
-        self.assertHasDisplayData(response, submission)
-
-    @patch('intake.notifications.slack_submissions_viewed.send')
-    def test_staff_user_can_get_submission_display(self, slack):
-        self.be_cfa_user()
-        submission = self.a_pubdef_submissions[0]
-        result = self.get_detail(submission)
-        self.assertEqual(result.context_data['form'].submission, submission)
-        self.assertHasDisplayData(result, submission)
-
-    @patch('intake.models.FillablePDF')
-    @patch('intake.notifications.slack_submissions_viewed.send')
-    def test_user_with_pdf_redirected_to_pdf(self, slack, FillablePDF):
-        self.be_sfpubdef_user()
-        submission = self.sf_pubdef_submissions[0]
-        result = self.get_detail(submission)
-        self.assertRedirects(
-            result,
-            reverse(
-                'intake-filled_pdf', kwargs=dict(submission_id=submission.id)),
-            fetch_redirect_response=False)
-        slack.assert_not_called()  # notification should be handled by pdf view
-        FillablePDF.assert_not_called()
-
-    @patch('intake.notifications.slack_submissions_viewed.send')
-    def test_user_cant_see_app_detail_for_other_county(self, slack):
-        self.be_ccpubdef_user()
-        submission = self.sf_pubdef_submissions[0]
-        response = self.get_detail(submission)
-        self.assertRedirects(response, reverse('intake-app_index'))
-        slack.assert_not_called()
-
-    @patch('intake.notifications.slack_submissions_viewed.send')
-    def test_user_can_see_app_detail_for_multi_county(self, slack):
-        self.be_apubdef_user()
-        submission = self.combo_submissions[0]
-        response = self.get_detail(submission)
-        self.assertHasDisplayData(response, submission)
-
-    @patch('intake.notifications.slack_submissions_viewed.send')
-    def test_agency_user_can_see_transfer_action_link(self, slack):
-        self.be_apubdef_user()
-        submission = self.a_pubdef_submissions[0]
-        response = self.get_detail(submission)
-        transfer_action_url = html_utils.conditional_escape(
-            submission.get_transfer_action(response.wsgi_request)['url'])
-        self.assertContains(response, transfer_action_url)
-
-    @patch('intake.notifications.slack_submissions_viewed.send')
-    def test_agency_user_can_see_case_printout_link(self, slack):
-        self.be_apubdef_user()
-        submission = self.a_pubdef_submissions[0]
-        response = self.get_detail(submission)
-        printout_url = html_utils.conditional_escape(
-            submission.get_case_printout_url())
-        self.assertContains(response, printout_url)
 
 
 class TestApplicationBundle(IntakeDataTestCase):
