+++ intake/views/app_detail_views.py
@@ -1,14 +1,12 @@
 from django.shortcuts import redirect
 from django.core.urlresolvers import reverse_lazy
-from django.views.generic import View
 from django.views.generic.base import TemplateView
 from django.contrib import messages
-from django.template.response import TemplateResponse
 
 import intake.services.submissions as SubmissionsService
-
 from intake import models, notifications
 from intake.views.base_views import ViewAppDetailsMixin
+
 
 NOT_ALLOWED_MESSAGE = str(
     "Sorry, you are not allowed to access that client information. "
@@ -21,14 +19,10 @@ def not_allowed(request):
     return redirect('intake-app_index')
 
 
-class ApplicationDetail(ViewAppDetailsMixin, View):
+class ApplicationDetail(ViewAppDetailsMixin, TemplateView):
     """Displays detailed information for an org user.
     """
     template_name = "app_detail.jinja"
-
-    def not_allowed(self, request):
-        messages.error(request, NOT_ALLOWED_MESSAGE)
-        return redirect('intake-app_index')
 
     def mark_viewed(self, request, submission):
         models.ApplicationLogEntry.log_opened([submission.id], request.user)
@@ -41,18 +35,30 @@ class ApplicationDetail(ViewAppDetailsMixin, View):
             return redirect(
                 reverse_lazy('intake-filled_pdf',
                              kwargs=dict(submission_id=submission_id)))
-        submissions = list(SubmissionsService.get_permitted_submissions(
+        self.submissions = list(SubmissionsService.get_permitted_submissions(
             request.user, [submission_id]))
-        if not submissions:
-            return self.not_allowed(request)
-        submission = submissions[0]
-        self.mark_viewed(request, submission)
+        if not self.submissions:
+            return not_allowed(request)
+        return super().get(request, submission_id)
+
+    def get_context_data(self, *args, **kwargs):
+        context = super().get_context_data(*args, **kwargs)
+        submission = self.submissions[0]
+        self.mark_viewed(self.request, submission)
         display_form, letter_display = submission.get_display_form_for_user(
-            request.user)
-        context = dict(
+            self.request.user)
+        context.update(
             form=display_form,
+            submission=submission,
             declaration_form=letter_display)
-        response = TemplateResponse(request, self.template_name, context)
-        return response
+        return context
+
+
+class ApplicationHistoryView(ApplicationDetail):
+    """Displays a list of information abotu the history of this application
+    """
+    template_name = "app_history.jinja"
+
 
 app_detail = ApplicationDetail.as_view()
+app_history = ApplicationHistoryView.as_view()
