+++ templates/app_history.jinja
@@ -1,0 +1,23 @@
+{% extends "app_detail.jinja" %}
+
+{%- block page_heading -%}
+Application History
+{%- endblock page_heading -%}
+
+
+{%- block app_detail_contents -%}
+<ul class="application-events-list">
+  <li class="application-event-block">
+    Something happened
+  </li>
+  <li class="application-event-block">
+    Another thing happened
+  </li>
+  <li class="application-event-block">
+    Temporal Anomaly!
+  </li>
+  <li class="application-event-block">
+    Nothing happened
+  </li>
+</ul>
+{%- endblock app_detail_contents -%}
